*: split src/progress.rs into mods
