+++ src/progress/progress_set.rs
@@ -11,48 +11,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Copyright 2015 The etcd Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 use eraftpb::{ConfState, SnapshotMetadata};
 use errors::{Error, Result};
 use hashbrown::hash_map::DefaultHashBuilder;
 use hashbrown::{HashMap, HashSet};
+use progress::progress::Progress;
 use std::cell::RefCell;
-use std::cmp;
 
 // Since it's an integer, it rounds for us.
 #[inline]
 fn majority(total: usize) -> usize {
     (total / 2) + 1
-}
-
-/// The state of the progress.
-#[derive(Debug, PartialEq, Clone, Copy)]
-pub enum ProgressState {
-    /// Whether it's probing.
-    Probe,
-    /// Whether it's replicating.
-    Replicate,
-    /// Whethers it's a snapshot.
-    Snapshot,
-}
-
-impl Default for ProgressState {
-    fn default() -> ProgressState {
-        ProgressState::Probe
-    }
 }
 
 /// A Raft internal representation of a Configuration.
@@ -670,434 +639,11 @@ impl ProgressSet {
     }
 }
 
-/// The progress of catching up from a restart.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Progress {
-    /// How much state is matched.
-    pub matched: u64,
-    /// The next index to apply
-    pub next_idx: u64,
-    /// When in ProgressStateProbe, leader sends at most one replication message
-    /// per heartbeat interval. It also probes actual progress of the follower.
-    ///
-    /// When in ProgressStateReplicate, leader optimistically increases next
-    /// to the latest entry sent after sending replication message. This is
-    /// an optimized state for fast replicating log entries to the follower.
-    ///
-    /// When in ProgressStateSnapshot, leader should have sent out snapshot
-    /// before and stop sending any replication message.
-    pub state: ProgressState,
-    /// Paused is used in ProgressStateProbe.
-    /// When Paused is true, raft should pause sending replication message to this peer.
-    pub paused: bool,
-    /// This field is used in ProgressStateSnapshot.
-    /// If there is a pending snapshot, the pendingSnapshot will be set to the
-    /// index of the snapshot. If pendingSnapshot is set, the replication process of
-    /// this Progress will be paused. raft will not resend snapshot until the pending one
-    /// is reported to be failed.
-    pub pending_snapshot: u64,
-
-    /// This is true if the progress is recently active. Receiving any messages
-    /// from the corresponding follower indicates the progress is active.
-    /// RecentActive can be reset to false after an election timeout.
-    pub recent_active: bool,
-
-    /// Inflights is a sliding window for the inflight messages.
-    /// When inflights is full, no more message should be sent.
-    /// When a leader sends out a message, the index of the last
-    /// entry should be added to inflights. The index MUST be added
-    /// into inflights in order.
-    /// When a leader receives a reply, the previous inflights should
-    /// be freed by calling inflights.freeTo.
-    pub ins: Inflights,
-}
-
-impl Progress {
-    /// Creates a new progress with the given settings.
-    pub fn new(next_idx: u64, ins_size: usize) -> Self {
-        Progress {
-            matched: 0,
-            next_idx,
-            state: ProgressState::default(),
-            paused: false,
-            pending_snapshot: 0,
-            recent_active: false,
-            ins: Inflights::new(ins_size),
-        }
-    }
-
-    fn reset_state(&mut self, state: ProgressState) {
-        self.paused = false;
-        self.pending_snapshot = 0;
-        self.state = state;
-        self.ins.reset();
-    }
-
-    pub(crate) fn reset(&mut self, next_idx: u64) {
-        self.matched = 0;
-        self.next_idx = next_idx;
-        self.state = ProgressState::default();
-        self.paused = false;
-        self.pending_snapshot = 0;
-        self.recent_active = false;
-        debug_assert!(self.ins.cap() != 0);
-        self.ins.reset();
-    }
-
-    /// Changes the progress to a probe.
-    pub fn become_probe(&mut self) {
-        // If the original state is ProgressStateSnapshot, progress knows that
-        // the pending snapshot has been sent to this peer successfully, then
-        // probes from pendingSnapshot + 1.
-        if self.state == ProgressState::Snapshot {
-            let pending_snapshot = self.pending_snapshot;
-            self.reset_state(ProgressState::Probe);
-            self.next_idx = cmp::max(self.matched + 1, pending_snapshot + 1);
-        } else {
-            self.reset_state(ProgressState::Probe);
-            self.next_idx = self.matched + 1;
-        }
-    }
-
-    /// Changes the progress to a Replicate.
-    pub fn become_replicate(&mut self) {
-        self.reset_state(ProgressState::Replicate);
-        self.next_idx = self.matched + 1;
-    }
-
-    /// Changes the progress to a snapshot.
-    pub fn become_snapshot(&mut self, snapshot_idx: u64) {
-        self.reset_state(ProgressState::Snapshot);
-        self.pending_snapshot = snapshot_idx;
-    }
-
-    /// Sets the snapshot to failure.
-    pub fn snapshot_failure(&mut self) {
-        self.pending_snapshot = 0;
-    }
-
-    /// Unsets pendingSnapshot if Match is equal or higher than
-    /// the pendingSnapshot
-    pub fn maybe_snapshot_abort(&self) -> bool {
-        self.state == ProgressState::Snapshot && self.matched >= self.pending_snapshot
-    }
-
-    /// Returns false if the given n index comes from an outdated message.
-    /// Otherwise it updates the progress and returns true.
-    pub fn maybe_update(&mut self, n: u64) -> bool {
-        let need_update = self.matched < n;
-        if need_update {
-            self.matched = n;
-            self.resume();
-        };
-
-        if self.next_idx < n + 1 {
-            self.next_idx = n + 1
-        }
-
-        need_update
-    }
-
-    /// Optimistically advance the index
-    pub fn optimistic_update(&mut self, n: u64) {
-        self.next_idx = n + 1;
-    }
-
-    /// Returns false if the given index comes from an out of order message.
-    /// Otherwise it decreases the progress next index to min(rejected, last)
-    /// and returns true.
-    pub fn maybe_decr_to(&mut self, rejected: u64, last: u64) -> bool {
-        if self.state == ProgressState::Replicate {
-            // the rejection must be stale if the progress has matched and "rejected"
-            // is smaller than "match".
-            if rejected <= self.matched {
-                return false;
-            }
-            self.next_idx = self.matched + 1;
-            return true;
-        }
-
-        // the rejection must be stale if "rejected" does not match next - 1
-        if self.next_idx == 0 || self.next_idx - 1 != rejected {
-            return false;
-        }
-
-        self.next_idx = cmp::min(rejected, last + 1);
-        if self.next_idx < 1 {
-            self.next_idx = 1;
-        }
-        self.resume();
-        true
-    }
-
-    /// Determine whether progress is paused.
-    pub fn is_paused(&self) -> bool {
-        match self.state {
-            ProgressState::Probe => self.paused,
-            ProgressState::Replicate => self.ins.full(),
-            ProgressState::Snapshot => true,
-        }
-    }
-
-    /// Resume progress
-    pub fn resume(&mut self) {
-        self.paused = false;
-    }
-
-    /// Pause progress.
-    pub fn pause(&mut self) {
-        self.paused = true;
-    }
-}
-
-/// A buffer of inflight messages.
-#[derive(Debug, PartialEq)]
-pub struct Inflights {
-    // the starting index in the buffer
-    start: usize,
-    // number of inflights in the buffer
-    count: usize,
-
-    // ring buffer
-    buffer: Vec<u64>,
-}
-
-// The `buffer` must have it's capacity set correctly on clone, normally it does not.
-impl Clone for Inflights {
-    fn clone(&self) -> Self {
-        let mut buffer = self.buffer.clone();
-        buffer.reserve(self.buffer.capacity() - self.buffer.len());
-        Inflights {
-            start: self.start,
-            count: self.count,
-            buffer,
-        }
-    }
-}
-
-impl Inflights {
-    /// Creates a new buffer for inflight messages.
-    pub fn new(cap: usize) -> Inflights {
-        Inflights {
-            buffer: Vec::with_capacity(cap),
-            start: 0,
-            count: 0,
-        }
-    }
-
-    /// Returns true if the inflights is full.
-    pub fn full(&self) -> bool {
-        self.count == self.cap()
-    }
-
-    /// The buffer capacity.
-    pub fn cap(&self) -> usize {
-        self.buffer.capacity()
-    }
-
-    /// Adds an inflight into inflights
-    pub fn add(&mut self, inflight: u64) {
-        if self.full() {
-            panic!("cannot add into a full inflights")
-        }
-
-        let mut next = self.start + self.count;
-        if next >= self.cap() {
-            next -= self.cap();
-        }
-        assert!(next <= self.buffer.len());
-        if next == self.buffer.len() {
-            self.buffer.push(inflight);
-        } else {
-            self.buffer[next] = inflight;
-        }
-        self.count += 1;
-    }
-
-    /// Frees the inflights smaller or equal to the given `to` flight.
-    pub fn free_to(&mut self, to: u64) {
-        if self.count == 0 || to < self.buffer[self.start] {
-            // out of the left side of the window
-            return;
-        }
-
-        let mut i = 0usize;
-        let mut idx = self.start;
-        while i < self.count {
-            if to < self.buffer[idx] {
-                // found the first large inflight
-                break;
-            }
-
-            // increase index and maybe rotate
-            idx += 1;
-            if idx >= self.cap() {
-                idx -= self.cap();
-            }
-
-            i += 1;
-        }
-
-        // free i inflights and set new start index
-        self.count -= i;
-        self.start = idx;
-    }
-
-    /// Frees the first buffer entry.
-    pub fn free_first_one(&mut self) {
-        let start = self.buffer[self.start];
-        self.free_to(start);
-    }
-
-    /// Frees all inflights.
-    pub fn reset(&mut self) {
-        self.count = 0;
-        self.start = 0;
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use harness::setup_for_test;
-    use progress::Inflights;
-
-    #[test]
-    fn test_inflight_add() {
-        setup_for_test();
-        let mut inflight = Inflights::new(10);
-        for i in 0..5 {
-            inflight.add(i);
-        }
-
-        let wantin = Inflights {
-            start: 0,
-            count: 5,
-            buffer: vec![0, 1, 2, 3, 4],
-        };
-
-        assert_eq!(inflight, wantin);
-
-        for i in 5..10 {
-            inflight.add(i);
-        }
-
-        let wantin2 = Inflights {
-            start: 0,
-            count: 10,
-            buffer: vec![0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
-        };
-
-        assert_eq!(inflight, wantin2);
-
-        let mut inflight2 = Inflights::new(10);
-        inflight2.start = 5;
-        inflight2.buffer.extend_from_slice(&[0, 0, 0, 0, 0]);
-
-        for i in 0..5 {
-            inflight2.add(i);
-        }
-
-        let wantin21 = Inflights {
-            start: 5,
-            count: 5,
-            buffer: vec![0, 0, 0, 0, 0, 0, 1, 2, 3, 4],
-        };
-
-        assert_eq!(inflight2, wantin21);
-
-        for i in 5..10 {
-            inflight2.add(i);
-        }
-
-        let wantin22 = Inflights {
-            start: 5,
-            count: 10,
-            buffer: vec![5, 6, 7, 8, 9, 0, 1, 2, 3, 4],
-        };
-
-        assert_eq!(inflight2, wantin22);
-    }
-
-    #[test]
-    fn test_inflight_free_to() {
-        setup_for_test();
-        let mut inflight = Inflights::new(10);
-        for i in 0..10 {
-            inflight.add(i);
-        }
-
-        inflight.free_to(4);
-
-        let wantin = Inflights {
-            start: 5,
-            count: 5,
-            buffer: vec![0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
-        };
-
-        assert_eq!(inflight, wantin);
-
-        inflight.free_to(8);
-
-        let wantin2 = Inflights {
-            start: 9,
-            count: 1,
-            buffer: vec![0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
-        };
-
-        assert_eq!(inflight, wantin2);
-
-        for i in 10..15 {
-            inflight.add(i);
-        }
-
-        inflight.free_to(12);
-
-        let wantin3 = Inflights {
-            start: 3,
-            count: 2,
-            buffer: vec![10, 11, 12, 13, 14, 5, 6, 7, 8, 9],
-        };
-
-        assert_eq!(inflight, wantin3);
-
-        inflight.free_to(14);
-
-        let wantin4 = Inflights {
-            start: 5,
-            count: 0,
-            buffer: vec![10, 11, 12, 13, 14, 5, 6, 7, 8, 9],
-        };
-
-        assert_eq!(inflight, wantin4);
-    }
-
-    #[test]
-    fn test_inflight_free_first_one() {
-        setup_for_test();
-        let mut inflight = Inflights::new(10);
-        for i in 0..10 {
-            inflight.add(i);
-        }
-
-        inflight.free_first_one();
-
-        let wantin = Inflights {
-            start: 1,
-            count: 9,
-            buffer: vec![0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
-        };
-
-        assert_eq!(inflight, wantin);
-    }
-}
-
-// TODO: Reorganize this whole file into separate files.
-// See https://github.com/pingcap/raft-rs/issues/125
 #[cfg(test)]
 mod test_progress_set {
+    use super::{Configuration, ProgressSet, Result};
     use hashbrown::HashSet;
-
-    use {progress::Configuration, Progress, ProgressSet, Result};
+    use progress::progress::Progress;
 
     const CANARY: u64 = 123;
